requirements.txt
Jinja2==3.0.0
Django==3.2.3
psycopg2-binary==2.8.6
djangorestframework==3.12.4
drf-spectacular==0.16.0
django-rest-auth==0.9.5
django-allauth==0.44.0
channels==3.0.3
channels-redis==3.2.0
celery==5.1.0
redis==3.5.3